Allow decimal values for margin rates on product registration
Update parsing logic and input field attributes in `registration.tsx` to support single decimal places for margin rates, replacing `parseInt` with `parseFloat` and adding `step="0.1"` to relevant input elements.

+++ client/src/pages/admin/products/registration.tsx
@@ -274,9 +274,9 @@ export default function ProductRegistrationPage() {
     if (bulkWrappingCost) data.wrappingCost = parseInt(bulkWrappingCost);
     if (bulkLaborCost) data.laborCost = parseInt(bulkLaborCost);
     if (bulkShippingCost) data.shippingCost = parseInt(bulkShippingCost);
-    if (bulkStartMargin) data.startMarginRate = parseInt(bulkStartMargin);
-    if (bulkDrivingMargin) data.drivingMarginRate = parseInt(bulkDrivingMargin);
-    if (bulkTopMargin) data.topMarginRate = parseInt(bulkTopMargin);
+    if (bulkStartMargin) data.startMarginRate = parseFloat(bulkStartMargin);
+    if (bulkDrivingMargin) data.drivingMarginRate = parseFloat(bulkDrivingMargin);
+    if (bulkTopMargin) data.topMarginRate = parseFloat(bulkTopMargin);
     
     if (Object.keys(data).length === 0) {
       toast({ variant: "destructive", title: "오류", description: "적용할 값을 입력해주세요" });
